requirements.txt
pytest>=7.4.0
playwright>=1.40.0
pytest-playwright>=0.3.0
pytest-rerunfailures>=12.0
allure-pytest>=2.13.2
PyYAML>=6.0
loguru>=0.7.0
